fix: stop peekBytes from truncating to 8 bits
std::min<uint8_t>(bytes, available) narrows both operands to uint8_t
before comparing, so any value above 255 wraps. An `available` of 256
yields 0 and the call returns an empty vector.

Both callers are diagnostics in ProtocolGame::parseMessage (the unknown
opcode warning and the parse exception handler), so their hex dumps are
silently truncated or empty exactly on the larger messages where they
are most useful.

Widening the comparison to int keeps the intended clamp: the result is
still bounded by `available`, so no additional bytes are ever read.

+++ src/framework/net/inputmessage.h
@@ -92,7 +92,9 @@ public:
         if (available <= 0)
             return {};
 
-        bytes = std::min<uint8_t>(bytes, available);
+        // std::min<uint8_t> would narrow both operands to uint8_t, so any value
+        // above 255 wraps: an `available` of 256 yields 0 bytes (empty result).
+        bytes = std::min<int>(bytes, available);
         std::vector<uint8_t> data(bytes);
         std::memcpy(data.data(), m_buffer + m_readPos, bytes);
         return data;
